Move wordbank manager to utils directory and add methods for counting number of unused words
